# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Alojamientos
  version: 1.0.0
  description: API para usuarios, anfitriones, alojamientos, reservas y comentarios. Rutas y respuestas basadas en los requerimientos del proyecto.
servers:
  - url: http://localhost:8080
paths:
  /usuarios:
    post:
      summary: Registrar un nuevo usuario
      operationId: registrarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroUsuario'
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Correo ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /usuarios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID del usuario
    get:
      summary: Obtener perfil de usuario
      operationId: obtenerUsuario
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Prohibido (sin permisos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Editar perfil de usuario
      operationId: editarUsuario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioEdicion'
      responses:
        '200':
          description: Perfil actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Prohibido (editar otro usuario)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegistroUsuario:
      type: object
      required:
        - nombre
        - email
        - contrasenia
        - telefono
        - fechaNacimiento
      properties:
        nombre:
          type: string
          example: "Carlos Pérez"
        email:
          type: string
          format: email
          example: "carlos@email.com"
        contrasenia:
          type: string
          example: "ContraseñaSegura1"
        telefono:
          type: string
          example: "3001234567"
        rol:
          type: string
          enum: [USUARIO, ANFITRION, ADMIN]
          example: "USUARIO"
        fechaNacimiento:
          type: string
          format: date
          example: "1990-10-10"
    UsuarioResponse:
      type: object
      properties:
        id:
          type: string
          example: "u_123456"
        nombre:
          type: string
          example: "Carlos Pérez"
        email:
          type: string
          format: email
        telefono:
          type: string
        rol:
          type: string
        fechaNacimiento:
          type: string
          format: date
    UsuarioEdicion:
      type: object
      properties:
        nombre:
          type: string
        telefono:
          type: string
        fotoPerfilUrl:
          type: string
        descripcion:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        codigo:
          type: string
          example: "ERR_VALIDACION"
        mensaje:
          type: string
          example: "El email ya está registrado"

    # -- Alojamiento --
...